Created new JSF where you handle accessToken, accessCode, message and number.
When you get redirected to home after login Java saves the accessCode
in the HomeController class. You can, in the same class, generate a
accessToken from the Utils class. This is where we will handle the
messages. The idea is to build a nice GUI to see conversations and to
be able to send SMS.